Rename compat.py to modules.py and update imports
